+++ CHANGELOG.md
@@ -13,6 +13,8 @@
   - `config.val`
     - Dropped support for `val` as a cofnig property, replaced by `valStrategy`
 - Dropped the `callback` propagator, to return later. It was all kinds of messed up, anyways. This propagator was the only anomaly in the code and now we can run with more assumptions.
+- Dropped support for custom `search` function option. May return later.
+- Dropped support for `initial_vars` in the `Solver` constructor options
 
 ## v2.3.4: